services/agent_api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent_api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "contracts",
  "skill_sdk",
  "router",
  "observability",
  "feedback",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.29.0",
  "pydantic-settings>=2.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
